requirements.txt
openinference-instrumentation-openai
openai
openinference-instrumentation-autogen
autogen
arize-otel
python-dotenv
psycopg2-binary

